Setup script improvements
* Update README.md to add sudoers rule for setup.sh, update setup.sh to
test password strength and to allow the script to be re-run as required
to complete the setup process including as the new user created during
the setup process so that setup.sh can then remove the setup user and
itself when it is finished.

* Update recipe to add setup.sh to the setup user's home directory and
also to /etc/skel.

* Update setup.sh, test hostname for standards compliance, improve
restartability, initial work on reencryption with a new volume key.

* Changes for reencryption.

* Reencryption fixes.

* Fixes for updating crypttab.

* Fix crypttab update, update README.md.

* Fix for enter_text function.

* Move setup.sh to /usr/local/bin, update the autostart script and sudoers
to match, update the recipe to match, add "Press enter to finish" to
setup.sh.

* Add instructions for making the disk image.

* Fix missing variable declaration.

* Rework completion check.

* Update links to the disk image, sha256sum and sha256sum.sig and bump
version to 0.9.1.

* Updates to README.md to add screenshot and to update formatting.

* Further updates to the readme files.

* Add information about connecting to wifi.

* Update setup.sh so that there is a reboot before the user logs in with
their new account. This is necessary because the setup user still owned
running processes otherwise and couldn't be removed as a result.